UI Improvments
Yeah

+++ packages/hardhat/contracts/TokenFactory.sol
@@ -16,10 +16,10 @@ contract TokenTemplate is ERC20 {
 
 contract TokenFactory {
 
-    event Create(address _Address, string name, string symbol, uint totalSupply);
+    event Create(address _Address, string name, string symbol, uint48 totalSupply);
     function create(uint totalSupply, string memory name, string memory symbol) public {
         TokenTemplate tokenTemplate = new TokenTemplate(totalSupply, name, symbol);
-        emit Create(address(tokenTemplate), name, symbol, totalSupply);
+        emit Create(address(tokenTemplate), name, symbol, uint48(totalSupply));
         console.log(totalSupply* 10**18, name, symbol);
     }
